+++ FuMonoStudios/Game/src/architecture/Component.h
@@ -1,7 +1,22 @@
 #pragma once
 #include "ecs.h"
+#include "Entity.h"
 
-class Component
-{
+class Manager;
+
+class Component {
+public:
+	Component() : ent_() {}
+	virtual ~Component() {}
+
+	inline void setContext(Entity* ent) {
+		ent_ = ent;
+	}
+
+	virtual void initComponent() {};
+	virtual void update() {};
+	virtual void render() {};
+protected:
+	Entity* ent_;
 };
