+++ FuMonoStudios/Game/src/architecture/ecs.h
@@ -6,6 +6,7 @@ class Entity;
 namespace ecs {
 
 	using cmpId_t = uint8_t;
+	//Namespace que contiene los IDs de todos los componentes
 	namespace cmp {
 		enum cmpId : cmpId_t {
 			TRANSFORM = 0,
@@ -16,7 +17,9 @@ namespace ecs {
 		};
 		constexpr cmpId_t maxComponentId = _LAST_CMP_ID;
 	}
+
 	using scId_t = uint8_t;
+	//Namespace para etiquetar las escenas
 	namespace sc {
 		enum sceneId : scId_t {
 			MAIN_SCENE,
@@ -24,24 +27,5 @@ namespace ecs {
 		};
 		constexpr sceneId maxSceneId = _LAST_SCENE;
 	}
-
-	class Component {
-	public:
-		Component() : ent_() {}
-		virtual ~Component() {}
-		inline void setContext(Entity* ent) {
-			ent_ = ent;
-		}
-
-		virtual void initComponent() {};
-		virtual void update() {};
-		virtual void render() {};
-	protected:
-		Entity* ent_;
-	};
-
-	class Manager {
-
-	};
 }
